Attempt #1 to fix connectivity problems:
After using an old computer for the first time in a while, I noticed it was having the connectivity problem.
So I checked the extension's console and saw that the response from mail.google.com was
"Failed to load resource: the server responded with a status of 401 (OK)"

After manually visiting mail.google.com, the problem went away. This mimic the workaround I tried.